Adding template configuration file support
If you have template variables that need to be re-used across multiple projects it's easier to to add these to a configuration file and link this across projects. This patch adds support for a configuration file "efc.handlebar.variables.json" to define the variables in this file. This is configuration in the constants.

I've implemented it as a service so it can be extended to add add other configuration values if needed.